various typos in the introduction chapter
One exception is that I have added FreeBSD to the list of supported
operating systems.  FreeBSD is the neglected server operating system.
With high profile Erlang applications, such as WhatsApp, running on
FreeBSD, it seems appropriate to mention it.

The discussion of Solaris does not mention that Oracle has ceased
development of it but that development is still going on with the
Illumos fork of OpenSolaris.  I would think that these facts would
also play into a decision about OS choice but I have not added them
since I did not want to overburden this pull request for typos with
additional discussion text.