typos in the processes chapter
I noticed some differences in the output of hipe_bifs/show_pcb/1 on
Erlang version 21.  The major differences seem to be with respect to
the various msg fields which have been moved out of the PCB.

I have not looked at the code changes in detail but it does look like
there will need to be some changes to the description of message handling.